Storefix (#21)
* fix: reset user store, add groups store

* merged master, updated qr components
